+++ src/main/java/chamapool/domain/chama/ChamaVO.java
@@ -1,10 +1,9 @@
 package chamapool.domain.chama;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record ChamaVO(
-    Integer chamaId, String name, LocalDate createdAt, BigDecimal accountBalance) {
+    Integer chamaId, String name, LocalDate createdAt, Double accountBalance) {
   public ChamaVO(Chama chama) {
     this(chama.chamaId(), chama.name(), chama.createdAt(), chama.accountBalance());
   }
